Builds working with some caveats
