+++ .github/workflows/release.yml
@@ -20,9 +20,17 @@ jobs:
         with:
           go-version: '1.16'
 
-      # TODO: for extra credit figure out how to add the supported SDK and tendermint versions
+      # TODO:figure out how to add the supported SDK and tendermint versions
+      # here we need to run the following and save the results to ENV in the goreleaser action
+      # SDKVERSION := $(shell go list -m -u -f '{{.Version}}' github.com/cosmos/cosmos-sdk)
+      # TMVERSION := $(shell go list -m -u -f '{{.Version}}' github.com/tendermint/tendermint)
+      # Also see .goreleaser.yaml to wire up the environment variables
+
+      # Once this is solved we should also be able to use https://github.com/pmorelli92/github-container-registry-build-push
+      # to build the right simd version and push it to githubs docker repo
+
       - run: echo ":rocket::rocket::rocket:" > ../release_notes.md 
-        # if: startsWith(github.ref, 'refs/tags/')
+        if: startsWith(github.ref, 'refs/tags/')
 
       - name: Build
         uses: goreleaser/goreleaser-action@v2
