+++ Makefile
@@ -1,4 +1,3 @@
-
 VERSION := $(shell echo $(shell git describe --tags) | sed 's/^v//')
 SDKVERSION := $(shell go list -m -u -f '{{.Version}}' github.com/cosmos/cosmos-sdk)
 TMVERSION := $(shell go list -m -u -f '{{.Version}}' github.com/tendermint/tendermint)
@@ -6,10 +5,10 @@ COMMIT  := $(shell git log -1 --format='%H')
 
 all: install
 
-LD_FLAGS = -X github.com/strangelove-ventures/horcrux/version.Version=$(VERSION) \
-	-X github.com/strangelove-ventures/horcrux/version.Commit=$(COMMIT) \
-	-X github.com/strangelove-ventures/horcrux/version.SDKVersion=$(SDKVERSION) \
-	-X github.com/strangelove-ventures/horcrux/version.TMVersion=$(TMVERSION)
+LD_FLAGS = -X github.com/strangelove-ventures/horcrux/cmd/horcrux/cmd.Version=$(VERSION) \
+	-X github.com/strangelove-ventures/horcrux/cmd/horcrux/cmd.Commit=$(COMMIT) \
+	-X github.com/strangelove-ventures/horcrux/cmd/horcrux/cmd.SDKVersion=$(SDKVERSION) \
+	-X github.com/strangelove-ventures/horcrux/cmd/horcrux/cmd.TMVersion=$(TMVERSION)
 
 BUILD_FLAGS := -ldflags '$(LD_FLAGS)'
 
@@ -29,9 +28,12 @@ clean:
 	rm -rf build
 
 build-simd-docker:
-	docker build -t jackzampolin/simd:$(SDKVERSION) -f ./docker/simd/Dockerfile ./docker/simd/
+	docker build -t jackzampolin/simd:$(SDKVERSION) --build-arg VERSION=$(SDKVERSION) -f ./docker/simd/Dockerfile ./docker/simd/
 
 build-horcrux-docker:
-	docker build -t strangelove-ventures/horcrux:$(VERSION) -f ./docker/horcrux/Dockerfile .
+	docker build -t jackzampolin/horcrux:$(VERSION) -f ./docker/horcrux/Dockerfile .
+
+push-simd-docker:
+	docker push jackzampolin/simd:$(SDKVERSION)
 
 .PHONY: all lint test race msan tools clean build
