+++ cmd/horcrux/cmd/config.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/mitchellh/go-homedir"
 	"github.com/strangelove-ventures/horcrux/signer"
-	"github.com/strangelove-ventures/horcrux/version"
 
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v2"
@@ -22,7 +21,6 @@ func init() {
 	// TODO: config peers add/remove
 	// TODO: config chain-id set
 	configCmd.AddCommand(initCmd())
-	rootCmd.AddCommand(version.NewVersionCommand())
 	rootCmd.AddCommand(configCmd)
 }
