+++ goreleaser.yaml
@@ -1,18 +1,22 @@
-project_name: sommelier
+project_name: horcrux
 
 env:
   # Require use of Go modules.
   - GO111MODULE=on
 
+# TODO: figure out how to embed the sdk and tendermint versions in here
+# https://goreleaser.com/customization/env/ set the env on the go releaser job
 builds:
-  - id: "sommelier"
-    main: ./cmd/sommelier/main.go
+  - id: "horcrux"
+    main: ./cmd/horcrux/main.go
     ldflags:
-      - -X github.com/strangelove-ventures/horcrux/version.Version={{ .Version }} -X github.com/strangelove-ventures/horcrux/version.Commit={{ .Commit }}
+      - -X github.com/strangelove-ventures/horcrux/version.Version={{ .Version }} \
+        -X github.com/strangelove-ventures/horcrux/version.Commit={{ .Commit }} \
+        -X github.com/strangelove-ventures/horcrux/version.SDKVersion=v0.43.2 \
+        -X github.com/strangelove-ventures/horcrux/version.TMVersion=v0.34.8
     goos:
       - darwin
       - linux
-      - windows
     goarch:
       - amd64
       - arm64
